Move navigation to presentation package.
