--- app/src/main/java/com/voltek/newsfeed/navigation/command/CommandOpenNewsSourcesScreen.kt
@@ -1,5 +1,0 @@
-package com.voltek.newsfeed.navigation.command
-
-import com.voltek.newsfeed.navigation.proxy.Command
-
-class CommandOpenNewsSourcesScreen : Command()
