--- app/src/main/java/com/voltek/newsfeed/navigation/command/CommandOpenWebsite.kt
@@ -1,8 +1,0 @@
-package com.voltek.newsfeed.navigation.command
-
-import com.voltek.newsfeed.navigation.proxy.Command
-
-/**
- * Open url via implicit intent
- */
-class CommandOpenWebsite(val url: String) : Command()
