--- app/src/main/java/com/voltek/newsfeed/navigation/command/CommandShareArticle.kt
@@ -1,5 +1,0 @@
-package com.voltek.newsfeed.navigation.command
-
-import com.voltek.newsfeed.navigation.proxy.Command
-
-class CommandShareArticle(val title: String, val url: String) : Command()
