--- app/src/main/java/com/voltek/newsfeed/navigation/command/CommandSystemMessage.kt
@@ -1,8 +1,0 @@
-package com.voltek.newsfeed.navigation.command
-
-import com.voltek.newsfeed.navigation.proxy.Command
-
-class CommandSystemMessage(
-        val message: String = "",
-        val error: Exception? = null
-) : Command()
