+++ app/src/main/java/com/voltek/newsfeed/presentation/navigation/command/CommandNavigatorAttached.kt
@@ -1,6 +1,6 @@
-package com.voltek.newsfeed.navigation.command
+package com.voltek.newsfeed.presentation.navigation.command
 
-import com.voltek.newsfeed.navigation.proxy.Command
+import com.voltek.newsfeed.presentation.navigation.proxy.Command
 
 /**
  * Placeholder command, that triggers queue execution when new navigator attached.
