+++ app/src/main/java/com/voltek/newsfeed/presentation/navigation/command/CommandOpenArticleDetailsScreen.kt
@@ -1,6 +1,6 @@
-package com.voltek.newsfeed.navigation.command
+package com.voltek.newsfeed.presentation.navigation.command
 
 import com.voltek.newsfeed.presentation.entity.ArticleUI
-import com.voltek.newsfeed.navigation.proxy.Command
+import com.voltek.newsfeed.presentation.navigation.proxy.Command
 
 class CommandOpenArticleDetailsScreen(val article: ArticleUI) : Command()
