+++ app/src/main/java/com/voltek/newsfeed/presentation/navigation/command/CommandOpenArticlesListScreen.kt
@@ -1,0 +1,5 @@
+package com.voltek.newsfeed.presentation.navigation.command
+
+import com.voltek.newsfeed.presentation.navigation.proxy.Command
+
+class CommandOpenArticlesListScreen : Command()
