+++ app/src/main/java/com/voltek/newsfeed/presentation/navigation/command/CommandOpenWebsite.kt
@@ -1,0 +1,8 @@
+package com.voltek.newsfeed.presentation.navigation.command
+
+import com.voltek.newsfeed.presentation.navigation.proxy.Command
+
+/**
+ * Open url via implicit intent
+ */
+class CommandOpenWebsite(val url: String) : Command()
