+++ app/src/main/java/com/voltek/newsfeed/presentation/ui/details/DetailsPresenter.kt
@@ -3,10 +3,10 @@ package com.voltek.newsfeed.presentation.ui.details
 import com.arellomobile.mvp.InjectViewState
 import com.voltek.newsfeed.NewsApp
 import com.voltek.newsfeed.presentation.entity.ArticleUI
-import com.voltek.newsfeed.navigation.command.CommandOpenNewsSourcesScreen
-import com.voltek.newsfeed.navigation.command.CommandOpenWebsite
-import com.voltek.newsfeed.navigation.command.CommandShareArticle
-import com.voltek.newsfeed.navigation.proxy.Router
+import com.voltek.newsfeed.presentation.navigation.command.CommandOpenNewsSourcesScreen
+import com.voltek.newsfeed.presentation.navigation.command.CommandOpenWebsite
+import com.voltek.newsfeed.presentation.navigation.command.CommandShareArticle
+import com.voltek.newsfeed.presentation.navigation.proxy.Router
 import com.voltek.newsfeed.presentation.base.BasePresenter
 import com.voltek.newsfeed.presentation.base.Event
 import com.voltek.newsfeed.presentation.ui.details.DetailsContract.DetailsModel
